#61 Extract PulseChannelCore from PulseChannel
